Remove the SSH pseudoterminal
This caused some issues with ffprobe if the remote command ever sends
something to stderr (which then gets sent to stdout locally, causing
format issues).

However, from some testing, this was original though needed to ensure
the 'q' sent to FFmpeg was sent directly over the pipe, but this seems
to not be required (any longer?).