Add state/process clear command
Helps simplify things in situations where a process terminates
unexpectedly and doesn't clean itself up. This can result in stale
process entries being left in the database. This command provides a
convenient way to clear such stuck process or state values, either for
the entire rffmpeg system, or for a particular host.