allow multiple ssh users
Different users on multiple servers were not possible because you had to set a specific user. I changed that so that setting no user in rffmpeg.yml and configuring the ssh config with a user works instead.

If you have configured  a specific user, this won't break your setup and disables the feature.